+++ src/main/java/com/softserveinc/ita/pageobjects/SearchResultPage.java
@@ -1,13 +1,22 @@
 package com.softserveinc.ita.pageobjects;
 
+import com.codeborne.selenide.CollectionCondition;
+import com.codeborne.selenide.SelenideElement;
+
 import java.util.List;
 
 import static com.codeborne.selenide.Selenide.$$x;
-import static com.softserveinc.ita.utils.runners.ElementsUtil.getListWithGoods;
+import static java.time.Duration.ofSeconds;
+import static java.util.stream.Collectors.toList;
 
 public class SearchResultPage extends BasePage<SearchResultPage> {
 
-    public List<String> getGoodsList(String item) {
-        return getListWithGoods($$x("//*[@class='goods-tile__title']"), item);
+    public List<String> getGoodsListBy(String productName) {
+        return $$x("//*[@class='goods-tile__title']")
+                .shouldBe(CollectionCondition.sizeGreaterThan(0), ofSeconds(8))
+                .stream()
+                .map(SelenideElement::getText)
+                .filter(text -> text.contains(productName))
+                .collect(toList());
     }
 }
