+++ src/test/java/com/softserveinc/ita/HomePageTest.java
@@ -1,6 +1,5 @@
 package com.softserveinc.ita;
 
-import com.softserveinc.ita.pageobjects.SearchResultPage;
 import com.softserveinc.ita.pageobjects.HomePage;
 import com.softserveinc.ita.pageobjects.ProductPage;
 import com.softserveinc.ita.utils.runners.TestRunner;
@@ -61,29 +60,5 @@ public class HomePageTest extends TestRunner {
         assertThat(isLanguageSwitched)
                 .as("Language should be switched")
                 .isTrue();
-    }
-
-    @Test
-    public void verifySearchResultsTest() {
-        var requiredItem = "Xiaomi Redmi Note 10";
-        homePage
-                .closeAdvertisingBannerIfDisplayed()
-                .setTextInSearchBar(requiredItem)
-                .performSearch();
-
-        var searchResultPage = new SearchResultPage();
-        var firstItem = searchResultPage.getGoodsList(requiredItem)
-                .stream()
-                .findFirst()
-                .toString();
-        
-        assertThat(firstItem)
-                .as("Test failed: The first item should contain: " + requiredItem)
-                .contains(requiredItem);
-
-        var lastItem = searchResultPage.getGoodsList(requiredItem);
-        assertThat(lastItem.get(lastItem.size() - 1))
-                .as("Test failed: The last item should contain: " + requiredItem)
-                .contains(requiredItem);
     }
 }
